% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/daily_obs.R
\name{daily_obs}
\alias{daily_obs}
\title{Get data from FMI (Finnish Meteorological Institute) stored queries.}
\usage{
daily_obs(key, fmisid, year)
}
\arguments{
\item{key}{FMI API-key}

\item{fmisid}{Weather station id (FMISID)}

\item{year}{Observation year}
}
\value{
data frame containing fmisid, name, time, tday, tmin, tmax, rain and snow columns
}
\description{
Get data from FMI (Finnish Meteorological Institute) stored queries.
}
\examples{
daily_obs('yourapikey', '100971', '2015')
}

